fix: Extract use*OnProxyStateChange hooks and fix Navigate conditions
